Move bundle logic out to be all in one place
